+++ src/components/Chart.jsx
@@ -22,13 +22,9 @@ const THEME = {
 };
 
 function getTitle(codeName, citations) {
-  /**
-   * Return title for chart.
-   */
   let reverseCitations = [];
   let yearBegin = 0;
   const minCitations = 10;
-  // go backwards; stop when hitting 10 citations or lower
   for (const point of citations.slice().reverse()) {
     if (point["y"] < minCitations) {
       break;
@@ -36,12 +32,9 @@ function getTitle(codeName, citations) {
     reverseCitations.push(point["y"]);
     yearBegin = point["x"];
   }
-
-  // Need at least two years with > minCitations
   if (reverseCitations.length < 2) {
     return codeName;
   }
-
   const ratio = reverseCitations[0] / reverseCitations.slice(-1)[0];
   const annualGrowth =
     (Math.pow(ratio, 1.0 / (reverseCitations.length - 1)) - 1) * 100;
@@ -51,59 +44,29 @@ function getTitle(codeName, citations) {
     annualGrowth.toFixed(1) +
     "% growth (year-over-year) since " +
     yearBegin;
-
   return title;
 }
 
 function SingleChart() {
-  /**
-   * Return chart for single code.
-   *
-   * Name of code is parsed from URI.
-   */
-
   let codeName = decodeURIComponent(useParams()["code"]);
   const series = getCodeCitations([codeName]);
-  const points =
-    Array.isArray(series) && series.length > 0 ? series[0].data : [];
+  const points = Array.isArray(series) && series.length > 0 ? series[0].data : [];
   const title = getTitle(codeName, points);
 
   const data = { id: codeName, data: points };
 
   return (
-    <Box
-      // note: for some reason, 100% has no effect
-      sx={{ width: "clamp(520px, 99%, 800px)" }}
-    >
-      <Paper
-        sx={{
-          p: 2,
-          display: "flex",
-          overflow: "auto",
-          flexDirection: "column",
-        }}
-      >
+    <Box sx={{ width: "clamp(520px, 99%, 800px)" }}>
+      <Paper sx={{ p: 2, display: "flex", overflow: "auto", flexDirection: "column" }}>
         {nivoChart([data], title, false, false, true)}
       </Paper>
     </Box>
   );
 }
 
-function nivoChart(
-  data,
-  title,
-  legend = true,
-  logScale = false,
-  clickable = false
-) {
-  /**
-   * Return nivo line-chart with default formatting for given data and title.
-   */
+function nivoChart(data, title, legend = true, logScale = false, clickable = false) {
   let legend_list = [];
   let margin_right = 50;
-
-  // For the MultiCodeChart page, we'll drop the legend even if legend=true is passed
-  // This is detected based on the logScale parameter which is only used for the trends page
   if (legend && !logScale) {
     legend_list = [
       {
@@ -121,25 +84,17 @@ function nivoChart(
         symbolShape: "circle",
         symbolBorderColor: "rgba(0, 0, 0, .5)",
         effects: [
-          {
-            on: "hover",
-            style: {
-              itemBackground: "rgba(0, 0, 0, .03)",
-              itemOpacity: 1,
-            },
-          },
+          { on: "hover", style: { itemBackground: "rgba(0, 0, 0, .03)", itemOpacity: 1 } },
         ],
       },
     ];
   }
 
-  // Validate data to ensure it contains points with valid coordinates
   const validData = data
     .map((series) => {
       if (!series.data || !Array.isArray(series.data)) {
         return { ...series, data: [] };
       }
-      // Filter out any data points with missing or invalid x or y values
       const validPoints = series.data
         .filter(
           (point) =>
@@ -156,83 +111,42 @@ function nivoChart(
     })
     .filter((series) => series.data.length > 0);
 
-  // Calculate max value for setting y-axis ticks and add 10% padding
   const maxYValue = validData.reduce((max, series) => {
     const seriesMax = series.data.reduce((m, point) => Math.max(m, point.y), 0);
     return Math.max(max, seriesMax);
   }, 0);
-
-  // Add 10% padding to max value to prevent cutting off
   const maxWithPadding = maxYValue * 1.1;
 
-  // Configure Y scale based on logScale parameter
   const yScale = logScale
-    ? {
-        type: "log",
-        base: 10,
-        min: LOG_Y_MIN, // Named constant for log scale minimum
-        max: maxWithPadding,
-        stacked: false,
-        reverse: false,
-      }
-    : {
-        type: "linear",
-        min: 0,
-        max: maxWithPadding,
-        stacked: false,
-        reverse: false,
-      };
+    ? { type: "log", base: 10, min: LOG_Y_MIN, max: maxWithPadding, stacked: false, reverse: false }
+    : { type: "linear", min: 0, max: maxWithPadding, stacked: false, reverse: false };
 
-  // Function to format numbers with apostrophes (e.g., 10'000)
-  const formatNumber = (value) => {
-    return value.toString().replace(/\B(?=(\d{3})+(?!\d))/g, "'");
-  };
+  const formatNumber = (value) => value.toString().replace(/\B(?=(\d{3})+(?!\d))/g, "'");
 
-  // Custom tick values for log scale to only show decades
   const getLogTickValues = (min, max) => {
     const tickValues = [];
     const minExp = Math.floor(Math.log10(min));
     const maxExp = Math.ceil(Math.log10(max));
-
     for (let i = minExp; i <= maxExp; i++) {
       const value = Math.pow(10, i);
-      if (value >= min) {
-        tickValues.push(value);
-      }
+      if (value >= min) tickValues.push(value);
     }
-
-    // Add LOG_Y_MIN as the first tick if it's our min value
-    if (min === LOG_Y_MIN && !tickValues.includes(LOG_Y_MIN)) {
-      tickValues.unshift(LOG_Y_MIN);
-    }
-
+    if (min === LOG_Y_MIN && !tickValues.includes(LOG_Y_MIN)) tickValues.unshift(LOG_Y_MIN);
     return tickValues;
   };
 
-  // If no valid data, display a message instead of an empty chart
   if (validData.length === 0) {
     return (
       <React.Fragment>
         <Title>{title}</Title>
-        <div
-          className="chart"
-          style={{
-            height: "400px",
-            display: "flex",
-            alignItems: "center",
-            justifyContent: "center",
-          }}
-        >
+        <div className="chart" style={{ height: "400px", display: "flex", alignItems: "center", justifyContent: "center" }}>
           <p>No valid data available for the selected options</p>
         </div>
       </React.Fragment>
     );
   }
 
-  // Get tick values for log scale
-  const tickValues = logScale
-    ? getLogTickValues(LOG_Y_MIN, maxWithPadding)
-    : undefined;
+  const tickValues = logScale ? getLogTickValues(LOG_Y_MIN, maxWithPadding) : undefined;
 
   return (
     <React.Fragment>
@@ -246,13 +160,7 @@ function nivoChart(
           yScale={yScale}
           axisTop={null}
           axisRight={null}
-          axisBottom={{
-            orient: "bottom",
-            legend: "Year",
-            legendOffset: 36,
-            legendPosition: "middle",
-            format: "d",
-          }}
+          axisBottom={{ orient: "bottom", legend: "Year", legendOffset: 36, legendPosition: "middle", format: "d" }}
           axisLeft={{
             orient: "left",
             legend: "Citations per year (Google Scholar)",
@@ -275,34 +183,9 @@ function nivoChart(
             const year = point.data.xFormatted ?? point.data.x;
             const value = point.data.yFormatted ?? point.data.y;
             return (
-              <div
-                style={{
-                  background: "#fff",
-                  padding: "6px 8px",
-                  border: `2px solid ${point.serieColor}`,
-                  borderRadius: 4,
-                  boxShadow: "0 2px 6px rgba(0,0,0,0.2)",
-                  color: "#333",
-                  fontSize: 12,
-                }}
-              >
-                <div
-                  style={{
-                    display: "flex",
-                    alignItems: "center",
-                    gap: 6,
-                    marginBottom: 4,
-                  }}
-                >
-                  <span
-                    style={{
-                      width: 10,
-                      height: 10,
-                      borderRadius: "50%",
-                      background: point.serieColor,
-                      display: "inline-block",
-                    }}
-                  />
+              <div style={{ background: "#fff", padding: "6px 8px", border: `2px solid ${point.serieColor}` , borderRadius: 4, boxShadow: "0 2px 6px rgba(0,0,0,0.2)", color: "#333", fontSize: 12 }}>
+                <div style={{ display: "flex", alignItems: "center", gap: 6, marginBottom: 4 }}>
+                  <span style={{ width: 10, height: 10, borderRadius: "50%", background: point.serieColor, display: "inline-block" }} />
                   <strong>{point.serieId}</strong>
                 </div>
                 <div>Year: {year}</div>
@@ -316,16 +199,11 @@ function nivoChart(
           theme={THEME}
           onClick={(point) => {
             if (!point || !point.data) return;
-            const year =
-              typeof point.data.x === "string"
-                ? parseInt(point.data.x, 10)
-                : point.data.x;
+            const year = typeof point.data.x === "string" ? parseInt(point.data.x, 10) : point.data.x;
             const codeName = point.serieId;
             const meta = codes[codeName];
             const url = buildScholarUrl(meta, year);
-            if (url) {
-              window.open(url, "_blank", "noopener,noreferrer");
-            }
+            if (url) window.open(url, "_blank", "noopener,noreferrer");
           }}
         />
       </div>
