--- src/components/index.js
@@ -1,1 +1,0 @@
-export { default as Chart } from "./Chart/chart";
