+++ src/components/index.jsx
@@ -1,0 +1,2 @@
+// Re-export chart helpers from the main Chart component file
+export { SingleChart, nivoChart } from "./Chart.jsx";
