--- src/index.js
@@ -1,22 +1,0 @@
-import React from "react";
-import { createRoot } from "react-dom/client";
-import { ThemeProvider, createTheme } from "@mui/material/styles";
-import reportWebVitals from "./reportWebVitals";
-import Dashboard from "./Dashboard";
-import "./style.css";
-
-const container = document.getElementById("root");
-const root = createRoot(container);
-const theme = createTheme();
-root.render(
-  <React.StrictMode>
-    <ThemeProvider theme={theme}>
-      <Dashboard />
-    </ThemeProvider>
-  </React.StrictMode>
-);
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
